removed extra containers
