--- src/components/Question.js
@@ -1,16 +1,0 @@
-import React from 'react'
-
-function Question(props) {
-
-  const { breed ,images } = props;
-
-  return (
-    <h2 className="question">{props.content}</h2>
-  )
-}
-
-// Question.propTypes = {
-//   content: React.PropTypes.string.isRequired
-// };
-
-export default Question
